+++ src/main/java/com/github/jlangch/venice/impl/docgen/DocGenerator.java
@@ -607,6 +607,7 @@ public class DocGenerator {
 		general.addItem(getDocItem("regex/pattern"));
 		general.addItem(getDocItem("regex/matcher"));
 		general.addItem(getDocItem("regex/find"));
+		general.addItem(getDocItem("regex/find-all"));
 		general.addItem(getDocItem("regex/find-group"));
 		general.addItem(getDocItem("regex/find-all-groups"));
 		general.addItem(getDocItem("regex/reset"));