+++ ChangeLog.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### added
 
 - added functions `regex/find-group` and `regex/find-all-groups`
+- added a completer for the REPL
 
 ### changes