+++ ChangeLog.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - added functions `regex/find-group` and `regex/find-all-groups`
 - added a completer for the REPL
 
+### fixed
+
+- fixed demo WEB application
+
 ### changes
 
 - added function `match?` that replaces `match`. `match` did not follow the naming conventions.